(GH-2755) Add support for running plans in noop mode
This adds basic support for running plans in noop mode. When running
`plan run --noop` or `Invoke-BoltPlan -Noop`, Bolt sets the executor
into noop mode. This automatically runs all tasks and apply blocks in a
plan in noop mode.

When running in noop mode, only a subset of plan functions are
supported. The plan functions that are not supported in noop mode are
`run_command`, `run_container`, `run_script`, `download_file`,
`upload_file`, `write_file`, and `file::write`, as they make changes to
either the controller or the specified targets and do not have a defined
noop mode. When any of these functions are encountered, the plan
immediately fails.

This also updates the `apply_prep` function to run the `puppet_library`
plugin in noop mode, which affects both the `apply --noop` command as
well as the `apply_prep` plan function. Because running this plugin in
noop mode might result in a target not having the puppet-agent package
installed, this also disables running the custom facts task in noop
mode, as the custom facts task depends on both the puppet-agent package
and Puppet's Ruby interpreter being available to the target.

!feature

* **Support running plans in noop mode**
  ([#2755](https://github.com/puppetlabs/bolt/issues/2755))

  Plans can now be run in no-operation mode from the command line using
  the `--noop` command-line option. Running in no-operation mode will
  automatically run all tasks and apply blocks in no-operation mode.
  Plans that use other plan functions to make changes to targets, such
  as `run_script`, will fail in no-operation mode, as these plan
  functions are not supported in this mode.

* **Run `puppet_agent::install` task in no-op mode**
  ([#2755](https://github.com/puppetlabs/bolt/issues/2755))

  The `puppet_agent::install` task can now be run in no-op mode.

* **Run plans in BoltSpec in no-op mode**
  ([#2755](https://github.com/puppetlabs/bolt/issues/2755))

  The BoltSpec `run_plan` function now supports running plans in no-op
  mode, allowing you to test the functionality of plans that support
  no-op mode.

+++ bolt-modules/file/lib/puppet/functions/file/write.rb
@@ -15,8 +15,15 @@ Puppet::Functions.create_function(:'file::write') do
   end
 
   def write(filename, content)
+    executor = Puppet.lookup(:bolt_executor) {}
+
+    # executor.noop is set when using 'plan run --noop' from the CLI
+    if executor&.noop
+      raise Bolt::Error.new('file::write is not supported in noop mode', 'bolt/noop-error')
+    end
+
     # Send Analytics Report
-    Puppet.lookup(:bolt_executor) {}&.report_function_call(self.class.name)
+    executor&.report_function_call(self.class.name)
 
     File.write(filename, content)
     nil
